requirements.txt
python-docx==1.1.0
flask==3.0.0
flask-cors==4.0.0
pytest==7.4.3
tabulate==0.9.0
python-dotenv==1.0.0
requests==2.31.0
